requirements.txt
DateTime==4.1.1
Django==1.9.5
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
MyApplication==0.1.0
Werkzeug==0.11.9
django-widget-tweaks==1.4.1
djangorestframework==3.4.1
itsdangerous==0.24
mysqlclient==1.3.7
pytz==2016.4
requests==2.10.0
zope.interface==4.1.3